Refactor update of build cache / version
