+++ website-react/package.json
@@ -15,7 +15,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "npm run insertBuildInfo; react-scripts build; node updateBuildCacheVersion.js; cp _redirects build/.",
+    "build": "react-scripts build; node updateBuildVersion.js; cp _redirects build/.",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "testNonInteractive": "react-scripts test --env=jsdom --coverage",
@@ -23,8 +23,7 @@
     "e2e": "cypress run",
     "e2eRecord": "cypress run --record",
     "ci": "start-server-and-test start http://localhost:3000 e2eRecord",
-    "lint": "standard --fix",
-    "insertBuildInfo": "replace './src/App.js' ${COMMIT_REF} ./src/App.js"
+    "lint": "standard --fix"
   },
   "devDependencies": {
     "cypress": "^3.0.1",
